Moved IcmpEchoHeader to icmp.rs
