+++ src/transport/mod.rs
@@ -1,3 +1,4 @@
+pub mod icmp;
 pub mod icmp4;
 pub mod icmp6;
 pub mod udp;
